web/requirements/dev.txt
-r base.txt

django-debug-toolbar>=6.0,<7
django-browser-reload>=1.19,<2